lint

+++ test/kubernetes/e2e/features/deployer/suite.go
@@ -298,7 +298,7 @@ func xdsClusterAssertion(testInstallation *e2e.TestInstallation) func(ctx contex
 				fmt.Sprintf("kgateway.%s.svc.cluster.local", testInstallation.Metadata.InstallNamespace),
 			), "xds socket address points to kgateway service, in installation namespace")
 
-			g.Expect(xdsSocketAddress.GetPortValue()).To(gomega.Equal(9977), "xds socket port points to kgateway service, in installation namespace")
+			g.Expect(xdsSocketAddress.GetPortValue()).To(gomega.Equal(uint32(9977)), "xds socket port points to kgateway service, in installation namespace")
 		}).
 			WithContext(ctx).
 			WithTimeout(time.Second * 10).
